# api.yml
# This is the OpenAPI specification for your service. It is used to generate the client code.
# To generate the server code, use the Gradle task `openApiGenerate`. The 
# generated code will be placed in `build/generate-resources/main`.
#
# We will evaluate you based on how well you design your API.
# 1. How well it follows REST principles.
# 2. How easy it is to understand and use.
#
# References
# 1. https://swagger.io/specification/
openapi: "3.0.0"
info:
  version: 1.0.0
  title: User Service
  license:
    name: MIT
servers:
  - url: http://localhost
paths:
  /estate:
    post:
      summary: Create a new estate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                width:
                  type: integer
                  minimum: 1
                  maximum: 50000
                length:
                  type: integer
                  minimum: 1
                  maximum: 50000
      responses:
        '201':
          description: Estate created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        '400':
          description: Invalid input
        '500':
          description: Internal server error
  /estate/{id}/tree:
    post:
      summary: Add a tree to an estate
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                x:
                  type: integer
                  minimum: 1
                y:
                  type: integer
                  minimum: 1
                height:
                  type: integer
                  minimum: 1
                  maximum: 30
      responses:
        '201':
          description: Tree added
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        '400':
          description: Invalid input
        '404':
          description: Estate not found
        '500':
          description: Internal server error
/estate/{id}/stats:
    get:
      summary: Get estate stats based on trees
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '200':
          description: Success get estate stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  max_height:
                    type: integer
                  min_height:
                    type: integer
                  median_height:
                    type: integer
        '404':
          description: Estate not found
        '500':
          description: Internal server error